50% restore LDULD/ MOVE function

+++ GPMVehicleControlSystem/Extension.cs
@@ -22,7 +22,7 @@ namespace GPMVehicleControlSystem
         }
         public static Quaternion ToQuaternion(this double Theta)
         {
-            double yaw_radians = Theta * Math.PI / 180.0;
+            double yaw_radians = (float)Theta * Math.PI / 180.0;
             double cos_yaw = Math.Cos(yaw_radians / 2.0);
             double sin_yaw = Math.Sin(yaw_radians / 2.0);
             return new Quaternion(0.0f, 0.0f, (float)sin_yaw, (float)cos_yaw);
